requirements.txt
PyKCS11==1.5.18
python_pkcs11==0.9.0
cryptography==46.0.1
lxml==6.0.1